backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-geojson==4.2.0
djangorestframework==3.15.2
GDAL==3.4.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
requests==2.32.3
djangorestframework-gis==1.1
